refactor: move registry classes to daemon-registry, accept List via constructor
Move LocalServiceMonitorRegistry, LocalServiceCollectorRegistry, and
LocalServiceDetectorRegistry from core/daemon-common to the new
core/daemon-registry module, rewriting them to accept their items
via constructor injection.

The new implementations drop ServiceLoader-based discovery and
reflection fallbacks. Callers will supply the exact list of monitors,
collectors, and detector factories they want to expose through
@Import'd Spring @Configuration classes in a follow-up task.

Until Task 5 replaces these with @Import, the existing caller beans
temporarily pass List.of() so the build still compiles.

DaemonProvisioningConfiguration's fallback ServiceDetectorRegistry
bean now uses an inline empty anonymous implementation, avoiding a
transitive ServiceMix spring-context 4.2.9 conflict that would
otherwise break SmartLifecycle resolution in daemon-common.
