+++ core/daemon-boot-collectd/pom.xml
@@ -182,6 +182,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.opennms.core</groupId>
+            <artifactId>org.opennms.core.daemon-registry</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- Kafka sink integration — for metric/event forwarding -->
         <dependency>
