+++ core/daemon-boot-minion/pom.xml
@@ -565,6 +565,13 @@
             </exclusions>
         </dependency>
 
+        <!-- LocalServiceDetectorRegistry (constructor-injected) -->
+        <dependency>
+            <groupId>org.opennms.core</groupId>
+            <artifactId>org.opennms.core.daemon-registry</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <!-- Concrete detector factory classes (exclusions match opennms-detectorclient-rpc) -->
         <dependency>
             <groupId>org.opennms</groupId>
