+++ core/daemon-boot-pollerd/pom.xml
@@ -182,6 +182,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <dependency>
+            <groupId>org.opennms.core</groupId>
+            <artifactId>org.opennms.core.daemon-registry</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- Poller implementation — Poller, QueryManagerDaoImpl, DefaultPollContext -->
         <dependency>
